Fix some 64 to 32 conversions for Windows.
One inconvenience is that it turns out the tellp() doesn't
actually return a numeric value, rather something that is
explicitly convertible to a numeric value. So we can't directly
apply the checked conversion macros to such results.

Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/88531
Commit-Queue: Tom Sepez <tsepez@chromium.org>

+++ core/fxge/win32/cpsoutput.cpp
@@ -25,11 +25,11 @@ bool CPSOutput::WriteBlock(const void* str, size_t len) {
 
     switch (m_mode) {
       case OutputMode::kExtEscape:
-        ExtEscape(m_hDC, PASSTHROUGH, send_len + 2,
+        ExtEscape(m_hDC, PASSTHROUGH, static_cast<int>(send_len + 2),
                   reinterpret_cast<const char*>(buffer), 0, nullptr);
         break;
       case OutputMode::kGdiComment:
-        GdiComment(m_hDC, send_len + 2, buffer);
+        GdiComment(m_hDC, static_cast<UINT>(send_len + 2), buffer);
         break;
     }
     input = input.subspan(send_len);
